+++ package.json
@@ -56,7 +56,7 @@
         "socket.io": "^2.0.3",
         "swagger-editor-dist": "3.6.0"
     },
-    "main": "./extension",
+    "main": "./src/extension",
     "contributes": {
         "commands": [
             {
